requirements.txt
Django~=2.2.12
python-environ==0.4.54
djangorestframework==3.12.2
psycopg2==2.8.6
gunicorn==20.0.4
